Avoid within-epsilon issues: change float types to double.
Now defines util::epsilon as 0.000000000000001
Remove weird floor in util::logistic(). Caused problem in Python, works here using double vs float.

+++ LocalSpace.h
@@ -93,7 +93,7 @@ public:
         assert (ls.is_orthonormal() && "handmade ls is orthonormal");
         assert (r.is_orthonormal() && "randomized ls is still orthonormal");
 
-        float e = util::epsilon * 10;
+        double e = util::epsilon * 10;
         assert (a.is_equal_within_epsilon(r.globalize(r.localize(a)), e));
         assert (a.is_equal_within_epsilon(r.localize(r.globalize(a)), e));
         assert (b.is_equal_within_epsilon(r.globalize(r.localize(b)), e));
